pyproject.toml
[project]
name = "jitapi"
version = "0.1.0"
description = "Just-in-Time API Orchestration System for LLMs - An MCP server for dynamic API discovery and execution"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["mcp", "api", "openapi", "llm", "orchestration", "claude", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "mcp>=1.0.0",
    "openapi-pydantic>=0.4.0",
    "pyyaml>=6.0",
    "chromadb>=0.4.0",
    "openai>=1.0.0",
    "networkx>=3.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
]

[project.scripts]
jitapi = "jitapi.main:main"

[project.urls]
Homepage = "https://github.com/nk3750/jitapi"
Documentation = "https://github.com/nk3750/jitapi#readme"
Repository = "https://github.com/nk3750/jitapi"
Issues = "https://github.com/nk3750/jitapi/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jitapi"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
